almost working...

+++ torchvision/ops/boxes.py
@@ -2,6 +2,7 @@ import torch
 
 
 def nms(boxes, scores, iou_threshold):
+    # type: (Tensor, Tensor, float)
     """
     Performs non-maximum suppression (NMS) on the boxes according
     to their intersection-over-union (IoU).
@@ -32,6 +33,7 @@ def nms(boxes, scores, iou_threshold):
 
 
 def batched_nms(boxes, scores, idxs, iou_threshold):
+    # type: (Tensor, Tensor, Tensor, float)
     """
     Performs non-maximum suppression in a batched fashion.
 
@@ -72,12 +74,13 @@ def batched_nms(boxes, scores, idxs, iou_threshold):
 
 
 def remove_small_boxes(boxes, min_size):
+    # type: (Tensor, float)
     """
     Remove boxes which contains at least one side smaller than min_size.
 
     Arguments:
         boxes (Tensor[N, 4]): boxes in (x1, y1, x2, y2) format
-        min_size (int): minimum size
+        min_size (float): minimum size
 
     Returns:
         keep (Tensor[K]): indices of the boxes that have both sides
@@ -90,6 +93,7 @@ def remove_small_boxes(boxes, min_size):
 
 
 def clip_boxes_to_image(boxes, size):
+    # type: (Tensor, Tuple[int, int])
     """
     Clip boxes so that they lie inside an image of size `size`.
